cleaned up redirects, navbar
