+++ src/router.js
@@ -9,6 +9,7 @@ import Commands from './views/commands.vue'
 const routes = [
     {path: '/', name:'landing', component: Landing},
     {path: '/callback', name:'auth', component: Landing},
+    {path: '/logout', name: 'logout', component: Landing},
     {path: '/commands', name:'commands', component: Commands},
     {path: '/servers', name:'servers', component: Servers},
     {path: '/servers/:key', name:'server', component: Server, props: true},
